Move development seeders to the development directory
