+++ package.json
@@ -117,8 +117,8 @@
     "db:seed:undo": "yarn build-seeders && sequelize db:seed:undo --seed ",
     "db:seed:all": "yarn build-seeders && sequelize db:seed:undo:all && sequelize db:seed:all",
     "db:seed:down:all": "yarn build-seeders && sequelize db:seed:undo:all",
-    "db:seed:development": "sequelize db:seed --seed $(ls dist_seeders/src/seeders/*.js)",
-    "db:seed:undo:development": "sequelize db:seed:undo:all --seeders-path ./dist_seeders/src/seeders",
+    "db:seed:development": "sequelize db:seed --seeders-path ./dist_seeders/src/seeders/development --seed $(ls dist_seeders/src/seeders/development/*.js)",
+    "db:seed:undo:development": "sequelize db:seed:undo:all --seeders-path ./dist_seeders/src/seeders/development",
     "db:seed:all:development": "yarn build-seeders && yarn db:seed:undo:development && yarn db:seed:development",
     "db:seed:production": "sequelize db:seed --seeders-path ./dist_seeders/src/seeders/production --seed $(ls dist_seeders/src/seeders/production/*.js)",
     "db:seed:undo:production": "sequelize db:seed:undo:all --seeders-path ./dist_seeders/src/seeders/production",
