+++ src/seeders/development/20200503190708-development-add-company-users.ts
@@ -1,14 +1,14 @@
 import { QueryInterface } from "sequelize";
-import { federico, aldana } from "./constants/admins";
-import { Environment } from "../config/Environment";
+import { claudio, marcos } from "../constants/companyUsers";
+import { Environment } from "../../config/Environment";
 
 export = {
   up: async (queryInterface: QueryInterface) => {
     if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
-    return queryInterface.bulkInsert("Admins", [federico.admin, aldana.admin]);
+    return queryInterface.bulkInsert("CompanyUsers", [claudio.companyUser, marcos.companyUser]);
   },
   down: async (queryInterface: QueryInterface) => {
     if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
-    return queryInterface.bulkDelete("Admins", {});
+    return queryInterface.bulkDelete("CompanyUsers", {});
   }
 };
