+++ src/seeders/development/20200503190712-development-add-applicants.ts
@@ -1,7 +1,7 @@
 import { QueryInterface } from "sequelize";
-import { dylan, manuel, sebastian } from "./constants/applicants";
-import { aldana } from "./constants/admins";
-import { Environment } from "../config/Environment";
+import { dylan, manuel, sebastian } from "../constants/applicants";
+import { aldana } from "../constants/admins";
+import { Environment } from "../../config/Environment";
 
 export = {
   up: async (queryInterface: QueryInterface) => {
