+++ src/seeders/development/20200612002940-development-add-admins.ts
@@ -1,14 +1,14 @@
 import { QueryInterface } from "sequelize";
-import { claudio, marcos } from "./constants/companyUsers";
-import { Environment } from "../config/Environment";
+import { federico, aldana } from "../constants/admins";
+import { Environment } from "../../config/Environment";
 
 export = {
   up: async (queryInterface: QueryInterface) => {
     if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
-    return queryInterface.bulkInsert("CompanyUsers", [claudio.companyUser, marcos.companyUser]);
+    return queryInterface.bulkInsert("Admins", [federico.admin, aldana.admin]);
   },
   down: async (queryInterface: QueryInterface) => {
     if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
-    return queryInterface.bulkDelete("CompanyUsers", {});
+    return queryInterface.bulkDelete("Admins", {});
   }
 };
