+++ src/seeders/development/20201226014356-development-add-shared-settings.ts
@@ -1,5 +1,5 @@
 import { QueryInterface } from "sequelize";
-import { Environment } from "../config/Environment";
+import { Environment } from "../../config/Environment";
 
 export = {
   up: async (queryInterface: QueryInterface) => {
